Revert "Lockscreen charging info: show decimal in battery temperature"
This reverts commit 4ad203dbcae4531c5ffbe1263139a50095c227a7.

+++ packages/SettingsLib/src/com/android/settingslib/fuelgauge/BatteryStatus.java
@@ -55,11 +55,11 @@ public class BatteryStatus {
     public final int maxChargingVoltage;
     public final int maxChargingWattage;
     public final boolean present;
-    public final float temperature;
+    public final int temperature;
 
     public BatteryStatus(int status, int level, int plugged, int health, boolean present,
             int maxChargingCurrent, int maxChargingVoltage,
-            int maxChargingWattage, float temperature) {
+            int maxChargingWattage, int temperature) {
         this.status = status;
         this.level = level;
         this.plugged = plugged;
